Created pursue move script
Trying to generalise the move script so that it can be mixed and matched with other components (like attacks).
Made a test enemy to try out the new script.
I need to fix how the move speed is calculated, currently accelerates too fast.
Also tweaked gravity and the throw arc.