Created enemy projectile attack script
Moved the old enemy attack in enemy behaviour to a new script. Attacks now begin being fired on a random delay.
Might like to make it so that enemies will only attack when they are able to see the player.
Also tested and fixed the wave manager, was ignoring the last object in the list.