Created a flame particle and invincibility frames
Flame will be used by a flamethrower enemy, and potentially for stage decoration/hazards.
As the flame produces a lot of particles, it when it collided with the player it destroyed them instantly.
I have wanted to add invincibility frames for a while, as enemies can hurt the player very quickly. There is now an adjustable timer set to PlayerHealth to manage the invincibility time.