Created a static, missile launching enemy
This enemy type was originally going to spawn flying enemies, but I thought the idea of spawning missiles made more sense and created a different kind of hazard to the flying enemies I already spawn.
Currently are only removed on collision with the floor or player, but having more interactions with the gun/throw might be interesting.