Fixed enemies being counted twice when dying
I noticed when playing, occasionally enemies would add double their usual score and waves would start earlier than intended.
This happened when the enemy was taking damage from 2 or more sources at the same time (for example, when an enemy is thrown at 2 enemies).
It was caused by a <= 0 check for the health script. When health was reduced from 2 sources, this check happened when health was equal to 0 and when it was -1.