Changed killfloor to bounce player instead of teleport
This change was made because of a bug, that when there are no enemies present the player would clip through the killfloor. I found through testing that the player was not being repositioned correctly though transform.position.
So I decided to add a force to the player instead. This actually feels a lot more fun, and the player can also bounce off of it without taking damage if they dash before colliding.

I also fixed the enemy count not resetting between plays.