Last minute fixes
Expanded sensitivity slider value to see if it works better on the webgl build.
Killfloor has been lowered, as some enemies collided with it on top of the arena. Flying enemy hover height increased.
Flipper enemy trigger increased.
Invincibility powerup time lengthened.
Changed the position of the throw to reduce the chance of enemies clipping through surfaces (it still seems to happen at some angles but otherwise is difficult to achieve).