Fixed the menu camera from moving away from the arena
The camera was not rotating, but moving while looking at the centre of the arena. It must have been making an oval shape around the arena, and therefore moving away slowly. This could be seen if the player left the menu screen on.
This was fixed by giving the camera a 'pivot' empty gameobject which it is a child of, and have it rotate instead.