Adding difficulty settings (no menu)
going to allow the player to set the speed of enemies and the amount of health the player has/heals.
allowing game speed to be reduced but not increased? making it faster felt weird, especially in combination with high enemy speed. may remove this if it is weird with slow speeds as well...